Move the gRPC Control impl to rl_control.rs
Rename only, plus the module path updates it needs to build. The next commit splits the Control service back out of it, so rl_control.rs reviews as a diff against the old control.rs.
